requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
allora_sdk==0.2.0
annotated-types==0.7.0
anyio==4.9.0
asn1crypto==1.5.1
attrs==25.3.0
bcl==2.3.1
bcrypt==4.3.0
bip-utils==2.9.3
bitarray==3.3.1
cachetools==5.5.0
cbor2==5.6.5
cdp-sdk==0.21.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
ckzg==2.1.1
click==8.1.8
coinbase-agentkit==0.4.0
coinbase-agentkit-openai-agents-sdk==0.3.0
coincurve==20.0.0
colorama==0.4.6
crcmod==1.7
cryptography==44.0.2
cytoolz==1.0.1
distlib==0.3.9
distro==1.9.0
ecdsa==0.19.1
ed25519-blake2b==1.4.1
egcd==2.0.2
eth-account==0.13.6
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.2.0
eth-utils==5.2.0
eth_abi==5.2.0
fastapi==0.115.12
filelock==3.16.1
frozenlist==1.5.0
griffe==1.7.2
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jiter==0.9.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
multidict==6.3.2
nilql==0.0.0a12
openai==1.70.0
openai-agents==0.0.6
packaging==24.2
pailliers==0.2.0
paramiko==3.5.1
parsimonious==0.10.0
platformdirs==4.3.6
pluggy==1.5.0
propcache==0.3.1
py-sr25519-bindings==0.2.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
PyNaCl==1.5.0
pyproject-api==1.8.0
pytest==8.3.5
pytest-asyncio==0.25.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyunormalize==16.0.0
rabinmiller==0.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
rpds-py==0.24.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
toolz==1.0.0
tox==4.23.2
tqdm==4.67.1
types-requests==2.32.0.20250328
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.28.1
web3==7.10.0
websockets==15.0.1
yarl==1.18.3
